requirements.txt


Django==2.2.4
djangorestframework==3.9.2
psycopg2-binary==2.8.4
django-extensions==2.2.5